+++ src/backend/sqlite/table.rs
@@ -212,6 +212,7 @@ impl StaticTypeMapper for DefaultTypeMapper {
                 ColumnType::Year(_) => unimplemented!("Year is not available in Sqlite."),
                 ColumnType::Bit(_) => unimplemented!("Bit is not available in Sqlite."),
                 ColumnType::VarBit(_) => unimplemented!("VarBit is not available in Sqlite."),
+                ColumnType::LTree => unimplemented!("LTree is not available in Sqlite."),
             }
         )
         .unwrap()